initial stable-23-3 sync

+++ contrib/restricted/boost/smart_ptr/include/boost/smart_ptr/detail/sp_counted_base_gcc_atomic.hpp
@@ -137,7 +137,7 @@ public:
 
     long use_count() const // nothrow
     {
-        return static_cast<long>( atomic_load( &use_count_ ) );
+        return atomic_load( &use_count_ );
     }
 };
